requirements.txt
# Macrova Backend — Sabitlenmiş bağımlılıklar
gunicorn==21.2.0
fastapi==0.111.0
uvicorn[standard]==0.30.1
supabase==2.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
python-dotenv==1.0.1
pydantic[email]==2.7.1
python-multipart==0.0.9
httpx==0.27.0
slowapi==0.1.9
bleach==6.3.0
cryptography==48.0.0
email-validator==2.3.0
